show summary on hover only when generated

+++ src/components/StagedVideoThumbnail/index.jsx
@@ -10,9 +10,12 @@ function StagedVideoThumbnail({ item }) {
 
     const { theme } = useContext(MainContext);
 
-    const renderTooltip = props => (
-        <Tooltip className='h-auto truncate tooltip' {...props}>{item?.metadata?.summary?.content?.split(' ').slice(0, 30).join(' ')}...</Tooltip>
-    );
+    const renderTooltip = props => {
+        if (item?.metadata?.summary?.content?.length > 0) {
+            return <Tooltip className='h-auto truncate tooltip' {...props}>{item?.metadata?.summary?.content?.split(' ').slice(0, 30).join(' ')}...</Tooltip>;
+        }
+        return <></>;
+    };
 
     return (
         <OverlayTrigger className='tooltip' placement="right" overlay={renderTooltip}>
